pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "grug"
version = "0.1.0"
description = "Self-hosted AI agent for TTRPG Discord communities"
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.3.0",
    "pydantic-ai-slim[anthropic,mcp]>=1.0.0",
    "anthropic>=0.80.0",
    "chromadb>=1.0.0",
    "APScheduler>=3.10.0",
    "SQLAlchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.0",
    "alembic>=1.13.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "aiofiles>=23.0.0",
    "httpx>=0.24.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
grug = "main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["grug*"]
